requirements.txt
pyserial>=3.5
pyautogui>=0.9.54
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
